Panel: Remove blueberry and mintupdate
- They look too small in Cinnamon (this needs fixing in the JS applet eventually),
or the size needs to be standardized.. as 24px icons actually look like 16px.

- Blueberry looked better in blue+white.

- The dot in mintupdate is nice but it doesn't work well in cinnamon because the
margin isn't standardized, so the icon doesn't look centered and can be too
close to adjacent ones.

We can consider re-adding these in the future, or maybe by then they'll use
an indicator fork and be DE-rendered. For now, let's stick to what we had
in the past.