Remove fullcolor monochrome actions
Apps need to switch to symbolic icons to achieve a nice
monochrome look, that's the only way that works well.

Failing, for apps which stick to fullcolor icons, we can't use mono,
since that can't work well with both light and dark themes, or themes
which use combinations of dark and light backgrounds.